fixed bug

+++ data/src/main/java/ar/cleaner/first/pf/data/repository_implementation/RamRepositoryImpl.kt
@@ -24,7 +24,7 @@ class RamRepositoryImpl @Inject constructor(
     override fun getAvailableRam(scope: CoroutineScope?): StateFlow<Double> =
         ramManager.getAvailableRam(scope)
 
-    override fun getTotalRam(): Double = ramManager.totalRam.toDouble()
+    override fun getTotalRam(): Double = ramManager.totalRam
 
     override suspend fun getBackgroundApps(): List<BackgroundApp> = ramManager.getBackgroundApps()
